move exception tests to separate tests file
